feat: minimal working test
