+++ .vscode-test.mjs
@@ -1,0 +1,6 @@
+import { defineConfig } from '@vscode/test-cli';
+
+export default defineConfig({
+	files: 'out/test/**/*.test.js',
+});
+
