+++ package.json
@@ -271,7 +271,7 @@
         "watch": "tsc -watch -p ./",
         "pretest": "npm run compile && npm run lint",
         "lint": "eslint src --ext ts",
-        "test": "node ./out/test/runTest.js"
+		"test": "rm -rf .vscode-test/user-data; vscode-test"
     },
     "devDependencies": {
         "@types/js-beautify": "^1.14.3",
@@ -280,8 +280,8 @@
         "@types/vscode": "^1.89.0",
         "@typescript-eslint/eslint-plugin": "^7.7.1",
         "@typescript-eslint/parser": "^7.7.1",
-        "@vscode/test-cli": "^0.0.9",
-        "@vscode/test-electron": "^2.4.0",
+		"@vscode/test-cli": "^0.0.8",
+		"@vscode/test-electron": "^2.3.9",
         "eslint": "^8.57.0"
     },
     "dependencies": {
